Move get-entity-from-data methods from space api to env api
