+++ lib/create-environment-api.js
@@ -898,7 +898,77 @@ export default function createEnvironmentApi ({
     }
   }
 
+  /**
+   * Creates SDK Entry object (locally) from entry data
+   * @memberof ContentfulSpaceAPI
+   * @param {object} entryData - Entry Data
+   * @return {Entry.Entry} Entry
+   * @example
+   * space.getEntry('entryId').then(entry => {
+   *
+   *   // Build a plainObject in order to make it usable for React (saving in state or redux)
+   *   const plainObject = entry.toPlainObject();
+   *
+   *   // The entry is being updated in some way as plainObject:
+   *   const updatedPlainObject = {
+   *     ...plainObject,
+   *     fields: {
+   *       ...plainObject.fields,
+   *       title: {
+   *         'en-US': 'updatedTitle'
+   *       }
+   *     }
+   *   };
+   *
+   *   // Rebuild an sdk object out of the updated plainObject:
+   *   const entryWithMethodsAgain = space.getEntryFromData(updatedPlainObject);
+   *
+   *   // Update with help of the sdk method:
+   *   entryWithMethodsAgain.update();
+   *
+   * });
+   **/
+  function getEntryFromData (entryData) {
+    return wrapEntry(http, entryData)
+  }
+
+  /**
+   * Creates SDK Asset object (locally) from entry data
+   * @memberof ContentfulSpaceAPI
+   * @param {object} entryData - Asset ID
+   * @return {Asset.Asset} Asset
+   * @example
+   * space.getAsset('asset_id').then(asset => {
+   *
+   *   // Build a plainObject in order to make it usable for React (saving in state or redux)
+   *   const plainObject = asset.toPlainObject();
+   *
+   *   // The asset is being updated in some way as plainObject:
+   *   const updatedPlainObject = {
+   *     ...plainObject,
+   *     fields: {
+   *       ...plainObject.fields,
+   *       title: {
+   *         'en-US': 'updatedTitle'
+   *       }
+   *     }
+   *   };
+   *
+   *   // Rebuild an sdk object out of the updated plainObject:
+   *   const assetWithMethodsAgain = space.getAssetFromData(updatedPlainObject);
+   *
+   *   // Update with help of the sdk method:
+   *   assetWithMethodsAgain.update();
+   *
+   * });
+  */
+  function getAssetFromData (assetData) {
+    return wrapAsset(http, assetData)
+  }
+
   return {
+    getEntryFromData,
+    getAssetFromData,
     delete: deleteEnvironment,
     update: updateEnvironment,
     getContentType: getContentType,
